+++ packages/near-membrane-base/src/__tests__/connector.spec.ts
@@ -2,16 +2,9 @@ import { createBlueConnector, createRedConnector } from '../../dist/index.mjs.js
 
 describe('createBlueConnector()', () => {
     it('throws when globalObject is not provided', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
-        // @ts-ignore
         expect(() => createBlueConnector()).toThrow();
     });
     it('returns connector function when globalObject is provided', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
         const connector = createBlueConnector(globalThis as any);
         expect(typeof connector).toBe('function');
     });
@@ -19,29 +12,16 @@ describe('createBlueConnector()', () => {
 
 describe('createRedConnector()', () => {
     it('throws when evaluator is not a function', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(4);
-        // @ts-ignore
         expect(() => createRedConnector()).toThrow();
-        // @ts-ignore
         expect(() => createRedConnector(null)).toThrow();
-        // @ts-ignore
         expect(() => createRedConnector(undefined)).toThrow();
-        // @ts-ignore
         expect(() => createRedConnector({})).toThrow();
     });
     it('returns connector function when requirements are satisfied', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
         const connector = createRedConnector(globalThis.eval);
         expect(typeof connector).toBe('function');
     });
     it('does not error in environments without globalThis', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(2);
         const globalThisRef = globalThis;
         const globalThisDesc = Reflect.getOwnPropertyDescriptor(globalThisRef, 'globalThis');
         Reflect.deleteProperty(globalThisRef, 'globalThis');
