+++ packages/near-membrane-node/src/__tests__/error.spec.js
@@ -23,39 +23,28 @@ const foo = {
 
 describe('The Error Boundary', () => {
     it('should preserve identity of errors after a membrane roundtrip', () => {
-        expect.assertions(3);
-
         const env = createVirtualEnvironment(globalThis, {
             endowments: Object.getOwnPropertyDescriptors({ foo }),
             globalObjectShape: globalThis,
         });
-
         env.evaluate(`foo.expose(() => { foo.a })`);
-
         expect(() => {
             sandboxedValue();
         }).toThrowError(Error);
-
         env.evaluate(`foo.expose(() => { foo.a = 1; })`);
-
         expect(() => {
             sandboxedValue();
         }).toThrowError(Error);
-
         env.evaluate(`foo.expose(() => { foo.b(2); })`);
-
         expect(() => {
             sandboxedValue();
         }).toThrowError(RangeError);
     });
     it('should remap the blue realm error instance to the sandbox errors', () => {
-        expect.assertions(3);
-
         const env = createVirtualEnvironment(globalThis, {
             endowments: Object.getOwnPropertyDescriptors({ expect, foo }),
             globalObjectShape: globalThis,
         });
-
         env.evaluate(`
             expect(() => {
                 foo.a;
@@ -73,19 +62,15 @@ describe('The Error Boundary', () => {
         `);
     });
     it('should capture throwing from user proxy', () => {
-        expect.assertions(3);
-
         const env = createVirtualEnvironment(globalThis, {
             endowments: Object.getOwnPropertyDescriptors({ foo }),
             globalObjectShape: globalThis,
         });
-
         env.evaluate(`
             const revocable = Proxy.revocable(() => undefined, {});
             revocable.revoke();
             foo.expose(revocable.proxy);
         `);
-
         expect(() => {
             // eslint-disable-next-line no-unused-expressions
             sandboxedValue.x;
@@ -102,7 +87,6 @@ describe('The Error Boundary', () => {
             endowments: Object.getOwnPropertyDescriptors({ foo }),
             globalObjectShape: globalThis,
         });
-
         expect(() => {
             env.evaluate(`
                 throw new TypeError('from sandbox');
@@ -114,7 +98,6 @@ describe('The Error Boundary', () => {
             endowments: Object.getOwnPropertyDescriptors({ foo }),
             globalObjectShape: globalThis,
         });
-
         expect(() => {
             env.evaluate(`
                 return; // illegal return statement
