+++ packages/near-membrane-shared/src/Function.ts
@@ -1,7 +1,3 @@
-export function identity<T>(value: T): T {
-    return value;
-}
-
 export function noop() {
     // No operation performed.
 }
