+++ test/dom/live-object.spec.js
