+++ test/environment/membrane.spec.js
@@ -2,23 +2,13 @@ import { createRedConnector } from '@locker/near-membrane-base';
 
 describe('createRedConnector()', () => {
     it('throws when evaluator is missing', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
-        // @ts-ignore
         expect(() => createRedConnector()).toThrow();
     });
     it('returns connector function when evaluator is present', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
         const connector = createRedConnector(window.eval);
         expect(typeof connector).toBe('function');
     });
     it('returns connector function for iframe.contentWindow', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
         const iframe = document.createElement('iframe');
         iframe.setAttribute('sandbox', 'allow-same-origin allow-scripts');
         document.body.appendChild(iframe);
@@ -27,9 +17,6 @@ describe('createRedConnector()', () => {
         expect(typeof connector).toBe('function');
     });
     it('returns connector function for iframe.contentWindow.window', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
         const iframe = document.createElement('iframe');
         iframe.setAttribute('sandbox', 'allow-same-origin allow-scripts');
         document.body.appendChild(iframe);
@@ -38,9 +25,6 @@ describe('createRedConnector()', () => {
         expect(typeof connector).toBe('function');
     });
     it('returns connector function for iframe.contentWindow.globalThis', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
         const iframe = document.createElement('iframe');
         iframe.setAttribute('sandbox', 'allow-same-origin allow-scripts');
         document.body.appendChild(iframe);
@@ -49,9 +33,6 @@ describe('createRedConnector()', () => {
         expect(typeof connector).toBe('function');
     });
     it('returns connector function for indirect ref eval', () => {
-        // Ignoring "Property 'assertions' does not exist on type '{...}'."
-        // @ts-ignore
-        expect.assertions(1);
         const iframe = document.createElement('iframe');
         iframe.setAttribute('sandbox', 'allow-same-origin allow-scripts');
         document.body.appendChild(iframe);
