+++ test/membrane/binary-data.spec.js
