+++ test/membrane/blue-expandos.spec.js
@@ -13,12 +13,9 @@ function saveFoo(arg) {
 
 describe('The blue expandos', () => {
     it('should never be subject to red side mutations', () => {
-        expect.assertions(1);
-
         const env = createVirtualEnvironment(window, {
             endowments: Object.getOwnPropertyDescriptors({ Base, saveFoo }),
         });
-
         env.evaluate(`
             function mixin(Clazz) {
                 return class extends Clazz {}
@@ -26,10 +23,8 @@ describe('The blue expandos', () => {
             const Foo = mixin(Base);
             saveFoo(Foo);
         `);
-
         class Test extends FooClazz {}
         const instance = new Test();
-
         expect(instance.statusVariable).toBe('initial');
     });
 });
