+++ test/membrane/cross-ns-extensions.spec.js
@@ -34,7 +34,6 @@ describe('The membrane', () => {
                 expect,
             }),
         });
-
         env.evaluate(`
             'use strict';
             expect(Foo.prototype.base()).toBe('from base');
