+++ test/membrane/freezing.spec.js
@@ -15,16 +15,13 @@ describe('Freezing', () => {
 
             window.plainObject = { x: 1 };
             Object.freeze(window.plainObject);
-
             const env = createVirtualEnvironment(window, {
                 globalObjectShape: window,
             });
-
             // Check the state of plainObject in the blue realm.
             expect(Object.isExtensible(window.plainObject)).toBe(false);
             expect(Object.isSealed(window.plainObject)).toBe(true);
             expect(Object.isFrozen(window.plainObject)).toBe(true);
-
             // Check the state of plainObject in the sandbox.
             env.evaluate(`
                 expect(Object.isExtensible(plainObject)).toBe(false);
@@ -39,7 +36,6 @@ describe('Freezing', () => {
                 }).toThrowError(TypeError);
                 expect(plainObject).toEqual({ x: 1 });
             `);
-
             expect(window.plainObject).toEqual({ x: 1 });
         });
         it('should be observed from within the sandbox (via endowments)', () => {
@@ -47,16 +43,13 @@ describe('Freezing', () => {
 
             const plainObject = { x: 1 };
             Object.freeze(plainObject);
-
             const env = createVirtualEnvironment(window, {
                 endowments: Object.getOwnPropertyDescriptors({ expect, plainObject }),
             });
-
             // Check the state of plainObject in the blue realm.
             expect(Object.isExtensible(plainObject)).toBe(false);
             expect(Object.isSealed(plainObject)).toBe(true);
             expect(Object.isFrozen(plainObject)).toBe(true);
-
             // Check the state of plainObject in the sandbox.
             env.evaluate(`
                 expect(Object.isExtensible(plainObject)).toBe(false);
@@ -71,7 +64,6 @@ describe('Freezing', () => {
                 }).toThrowError(TypeError);
                 expect(plainObject).toEqual({ x: 1 });
             `);
-
             expect(plainObject).toEqual({ x: 1 });
         });
         it('should be observed after passing an object back and forth', () => {
@@ -80,11 +72,9 @@ describe('Freezing', () => {
             const receiveX = {
                 ref: null,
             };
-
             const transmitX = {
                 ref: null,
             };
-
             const env = createVirtualEnvironment(window, {
                 endowments: Object.getOwnPropertyDescriptors({ receiveX, transmitX }),
                 globalObjectShape: window,
@@ -92,17 +82,14 @@ describe('Freezing', () => {
                     return true;
                 },
             });
-
             env.evaluate(`
                 expect(transmitX.ref).toBe(null);
                 expect(receiveX.ref).toBe(null);
             `);
-
             // Set the value of transmitX.ref to a POJO.
             transmitX.ref = {
                 x: 1,
             };
-
             // Verify that the POJO of transmitX.ref is reflected in the sandbox.
             // Set the value of receiveX.ref to the value of transmitX.ref WITHIN
             // the sandbox.
@@ -110,7 +97,6 @@ describe('Freezing', () => {
                 expect(transmitX.ref.x).toBe(1);
                 receiveX.ref = transmitX.ref;
             `);
-
             // Verify that the value of receiveX.ref is reflected in system mode.
             expect(receiveX.ref).toBe(transmitX.ref);
             Object.freeze(receiveX.ref);
@@ -122,7 +108,6 @@ describe('Freezing', () => {
             }).toThrowError(TypeError);
             expect(receiveX.ref.x).toBe(1);
             expect(transmitX.ref.x).toBe(1);
-
             // Return to the sandbox and verify that ref is frozen.
             env.evaluate(`
                 'use strict';
@@ -139,11 +124,9 @@ describe('Freezing', () => {
             expect.assertions(9);
 
             window.exoticObject = new ExoticObject({ x: 1 });
-
             const env = createVirtualEnvironment(window, {
                 globalObjectShape: window,
             });
-
             // Check the state of exoticObject in the sandbox.
             env.evaluate(`
                 expect(Object.isExtensible(window.exoticObject)).toBe(true);
@@ -151,13 +134,11 @@ describe('Freezing', () => {
                 expect(Object.isFrozen(window.exoticObject)).toBe(false);
                 exoticObject.y = 2; // Mutation makes the exotic object proxy static.
             `);
-
             // Freeze blue exoticObject after being observed by the sandbox.
             Object.freeze(window.exoticObject);
             expect(Object.isExtensible(window.exoticObject)).toBe(false);
             expect(Object.isSealed(window.exoticObject)).toBe(true);
             expect(Object.isFrozen(window.exoticObject)).toBe(true);
-
             // Verify the state of red exoticObject from within the sandbox.
             env.evaluate(`
                 'use strict';
@@ -166,7 +147,6 @@ describe('Freezing', () => {
                 }).not.toThrowError(TypeError);
                 expect({ ...exoticObject }).toEqual({ x: 1, y: 2, z: 3 });
             `);
-
             // Verify the sandboxed expando doesn't leak to blue exoticObject.
             expect({ ...window.exoticObject }).toEqual({ x: 1 });
         });
@@ -174,14 +154,12 @@ describe('Freezing', () => {
             expect.assertions(9);
 
             window.plainObject = { x: 1 };
-
             const env = createVirtualEnvironment(window, {
                 globalObjectShape: window,
                 liveTargetCallback() {
                     return true;
                 },
             });
-
             // Check the state of plainObject in the sandbox.
             env.evaluate(`
                 expect(Object.isExtensible(window.plainObject)).toBe(true);
@@ -189,13 +167,11 @@ describe('Freezing', () => {
                 expect(Object.isFrozen(window.plainObject)).toBe(false);
                 plainObject.y = 2; // Mutation makes the POJO proxy live.
             `);
-
             // Freeze blue plainObject after being observed by the sandbox.
             Object.freeze(window.plainObject);
             expect(Object.isExtensible(window.plainObject)).toBe(false);
             expect(Object.isSealed(window.plainObject)).toBe(true);
             expect(Object.isFrozen(window.plainObject)).toBe(true);
-
             // Verify the state of red plainObject from within the sandbox.
             env.evaluate(`
                 'use strict';
@@ -204,7 +180,6 @@ describe('Freezing', () => {
                 }).toThrowError(TypeError);
                 expect(plainObject).toEqual({ x: 1, y: 2 });
             `);
-
             // Verify the state of blue plainObject.
             expect(window.plainObject).toEqual({ x: 1, y: 2 });
         });
@@ -228,11 +203,9 @@ describe('Freezing', () => {
                     func.y = 2;
                 }).toThrowError(TypeError);
             };
-
             const env = createVirtualEnvironment(window, {
                 globalObjectShape: window,
             });
-
             env.evaluate(`
                 'use strict';
                 const object = { x: 1 };
@@ -258,14 +231,12 @@ describe('Freezing', () => {
                     return false;
                 },
             });
-
             const env = createVirtualEnvironment(window, {
                 endowments: Object.getOwnPropertyDescriptors({
                     exoticObject,
                     expect,
                 }),
             });
-
             env.evaluate(`
                 'use strict';
                 expect(() => {
@@ -297,7 +268,6 @@ describe('Freezing', () => {
                     return true;
                 },
             });
-
             env.evaluate(`
                 'use strict';
                 expect(() => {
